sistemato check_and_print_similarity, un bug in create_triplets, aggiunta una variabile globale per i threads in OMP, allineate le chiamate per misurare i tempi

+++ pthread_prod_cons/get_similarities.h
@@ -13,15 +13,15 @@
 */
 struct sign_doc{
     long long unsigned signature;
-    long long unsigned doc_id;
+    int doc_id;
 };
 
 /*
     per la coppia di documenti, la shared_signatures indica quante signatures hanno in comune. Se questa coppia esiste, il valore minimo è 1.
 */
 struct doc_couple{
-    long long unsigned doc_id;
-    long long unsigned doc2_id;
+    int doc_id;
+    int doc2_id;
     int shared_signatures;
 };
